# docs/api/openapi.yaml
...
openapi: "3.1.0"
info:
  title: sisys API
  description: AI 驱动的战略规划与决策智能平台 API
  version: "0.1.0"
  contact:
    name: sisys Team
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
paths:
...
  /auth/login:
    post:
      tags:
        - authentication
      summary: 用户登录
      description: 验证用户凭证并返回 JWT 令牌
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        "401":
          description: 无效凭证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "423":
          description: 账户已锁定
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 100
          description: 用户名
        password:
          type: string
          minLength: 1
          description: 密码
...
    TokenResponse:
      type: object
      required:
        - access_token
        - token_type
        - expires_in
        - user
      properties:
        access_token:
          type: string
          description: JWT access token
        refresh_token:
          type: string
          nullable: true
          description: Refresh token（可选）
        token_type:
          type: string
          default: bearer
          description: Token 类型
        expires_in:
          type: integer
          description: 过期时间（秒）
        user:
          $ref: '#/components/schemas/UserResponse'
    UserResponse:
      type: object
      required:
        - id
        - username
        - roles
      properties:
        id:
          type: string
          description: 用户 UUID
        username:
          type: string
          description: 用户名
        roles:
          type: array
          items:
            type: string
          description: 角色列表

    # --------------------------------------------------------------------------
    # 角色相关 schemas
    # --------------------------------------------------------------------------
...
    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: 错误详情

    # --------------------------------------------------------------------------
    # Embedding API schemas (Story 3-1a)
    # --------------------------------------------------------------------------
...
  securitySchemes:
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://auth.example.com/authorize
          tokenUrl: https://auth.example.com/token
          scopes: {}
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token 认证